C++ Constructors parameters
The following class have brand, model and year attributes, and a constructor with different parameters. Inside the constructor, we set the attributes equal to the constructor parameters (brand=x, etc.). When we call the constructor (by creating an object of the class), we pass parameters to the constructor, which will set the value of the corresponding attributes to the same: